# docs/openapi/rbac.yaml
openapi: 3.1.0
info:
  title: HandLab API - RBAC (Users, Organizations, Roles, Permissions, Memberships)
  description: |
    Contrato de API para o módulo RBAC (Role-Based Access Control) do sistema HandLab.
    
    ## Recursos
    - **Users**: Cadastro de usuários do sistema
    - **Organizations**: Organizações/clubes (V1: clube único - R34)
    - **Roles**: Catálogo de papéis (superadmin, dirigente, coordenador, treinador, atleta)
    - **Permissions**: Catálogo de permissões por papel
    - **Memberships**: Vínculos user↔organization+role
    
    ## Regras de Negócio
    - R25/R26: Permissões por papel e escopo
    - R29/R33: Exclusão lógica e histórico com rastro
    - R31/R32: Ações críticas auditadas
    - R34: Clube único na V1 (contexto organizacional obrigatório)
    - R42: Usuários sem vínculo ativo não podem operar
    
    ## Autenticação
    Todos os endpoints requerem autenticação via JWT Bearer token.
  version: 1.0.0
  contact:
    name: HandLab Team
  license:
    name: Proprietary
servers:
  - url: /v1
    description: API v1
security:
  - bearerAuth: []
...
paths:
...
  /organizations:
    get:
      operationId: listOrganizations
      tags: [organizations]
      summary: Listar organizações (paginado)
      description: |
        Lista paginada de organizações/clubes.
        
        **Nota V1:** Clube único (R34) — contexto organizacional implícito no token.
        
        **Regras:** R34 (clube único), R25/R26 (permissões)
      x-rule-ids: [R34, R25, R26]
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
        - name: order_by
          in: query
          schema:
            type: string
            enum: [created_at, name, updated_at]
            default: created_at
        - name: order_dir
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: search
          in: query
          description: Busca por name ou code
          schema:
            type: string
      responses:
        '200':
          description: Lista de organizações
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationPaginatedResponse'
              example:
                items:
                  - id: "660e8400-e29b-41d4-a716-446655440001"
                    name: "Clube Handebol São Paulo"
                    code: "CHSP"
                    created_at: "2024-01-01T00:00:00Z"
                    updated_at: "2024-01-01T00:00:00Z"
                page: 1
                limit: 50
                total: 1
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

    post:
      operationId: createOrganization
      tags: [organizations]
      summary: Criar organização
      description: |
        Cria uma nova organização/clube.
        
        **Regras:** R25/R26 (permissões), R29 (exclusão lógica), R31/R32 (auditoria)
      x-rule-ids: [R25, R26, R29, R31, R32]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationCreate'
            example:
              name: "Clube Handebol São Paulo"
              code: "CHSP"
      responses:
        '201':
          description: Organização criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          description: Conflito de unicidade (code/name duplicado)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: conflict_unique
                message: "Código ou nome já cadastrado"
                details:
                  constraint: "organizations_code_key"
        '422':
          $ref: '#/components/responses/ValidationError'
  /organizations/{organization_id}:
    get:
      operationId: getOrganizationById
      tags: [organizations]
      summary: Obter organização por ID
      description: |
        Retorna os dados de uma organização específica.
        
        **Regras:** R25/R26 (permissões)
      x-rule-ids: [R25, R26]
      parameters:
        - $ref: '#/components/parameters/organizationIdParam'
      responses:
        '200':
          description: Dados da organização
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    patch:
      operationId: updateOrganization
      tags: [organizations]
      summary: Atualizar organização
      description: |
        Atualiza dados de uma organização existente.
        
        **Regras:** R25/R26 (permissões), R31/R32 (auditoria)
      x-rule-ids: [R25, R26, R31, R32]
      parameters:
        - $ref: '#/components/parameters/organizationIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationUpdate'
            example:
              name: "Clube Handebol SP - Atualizado"
      responses:
        '200':
          description: Organização atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          description: Conflito de unicidade
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: conflict_unique
                message: "Código ou nome já cadastrado"
                details:
                  constraint: "organizations_code_key"
        '422':
          $ref: '#/components/responses/ValidationError'

  # ===========================================================================
  # ROLES (Catálogo - somente leitura)
  # ===========================================================================
...
  /permissions:
    get:
      operationId: listPermissions
      tags: [permissions]
      summary: Listar permissões disponíveis
      description: |
        Lista as permissões disponíveis no sistema (catálogo).
        Opcionalmente filtra por papel (role_code).
        
        **Regras:** R25/R26 (permissões)
      x-rule-ids: [R25, R26]
      parameters:
        - name: role_code
          in: query
          description: Filtrar permissões por papel
          schema:
            type: string
            enum: [superadmin, dirigente, coordenador, treinador, atleta]
      responses:
        '200':
          description: Lista de permissões
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Permission'
              example:
                - code: read_athlete
                  name: "Visualizar Atletas"
                  description: "Permissão para visualizar dados de atletas"
                  roles: [superadmin, dirigente, coordenador, treinador, atleta]
                - code: edit_athlete
                  name: "Editar Atletas"
                  description: "Permissão para editar dados de atletas"
                  roles: [superadmin, dirigente, coordenador, treinador]
                - code: admin_memberships
                  name: "Administrar Vínculos"
                  description: "Permissão para criar/editar vínculos de usuários"
                  roles: [superadmin, dirigente]
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /permissions/{permission_code}:
    get:
      operationId: getPermissionByCode
      tags: [permissions]
      summary: Obter permissão por código
      description: |
        Retorna os dados de uma permissão específica.
        
        **Regras:** R25/R26 (permissões)
      x-rule-ids: [R25, R26]
      parameters:
        - $ref: '#/components/parameters/permissionCodeParam'
      responses:
        '200':
          description: Dados da permissão
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Permission'
              example:
                code: edit_training
                name: "Editar Treinos"
                description: "Permissão para criar/editar sessões de treino"
                roles: [superadmin, dirigente, coordenador, treinador]
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # ===========================================================================
  # MEMBERSHIPS (Vínculos user↔organization+role)
  # ===========================================================================
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido via autenticação

  # ===========================================================================
  # Parameters
  # ===========================================================================
  parameters:
    pageParam:
      name: page
      in: query
      description: Número da página (1-based)
      schema:
        type: integer
        minimum: 1
        default: 1
    limitParam:
      name: limit
      in: query
      description: Itens por página (máx 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
...
    organizationIdParam:
      name: organization_id
      in: path
      required: true
      description: ID único da organização
      schema:
        type: string
        format: uuid
...
    permissionCodeParam:
      name: permission_code
      in: path
      required: true
      description: Código da permissão
      schema:
        type: string

  # ===========================================================================
  # Schemas
  # ===========================================================================
  schemas:
...
    Organization:
      type: object
      description: Organização/clube
      required:
        - id
        - name
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
          description: ID único da organização
        name:
          type: string
          minLength: 2
          maxLength: 255
          description: Nome da organização
        code:
          type: string
          nullable: true
          maxLength: 50
          description: Código curto (sigla)
        deleted_at:
          type: string
          format: date-time
          nullable: true
          readOnly: true
          description: Data/hora de exclusão lógica (RDB4)
        deleted_reason:
          type: string
          nullable: true
          readOnly: true
          description: Motivo da exclusão lógica
        created_at:
          type: string
          format: date-time
          readOnly: true
          description: Data/hora de criação
        updated_at:
          type: string
          format: date-time
          readOnly: true
          description: Data/hora da última atualização
    OrganizationCreate:
      type: object
      description: Payload para criação de organização
      required:
        - name
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 255
          description: Nome da organização
        code:
          type: string
          nullable: true
          maxLength: 50
          description: Código curto (sigla)
    OrganizationUpdate:
      type: object
      description: Payload para atualização de organização
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 255
          description: Nome da organização
        code:
          type: string
          nullable: true
          maxLength: 50
          description: Código curto (sigla)
    OrganizationPaginatedResponse:
      type: object
      description: Resposta paginada de organizações
      required:
        - items
        - page
        - limit
        - total
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Organization'
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          minimum: 0

    # -------------------------------------------------------------------------
    # Role Schema
    # -------------------------------------------------------------------------
...
    Permission:
      type: object
      description: Permissão do sistema (catálogo)
      required:
        - code
        - name
        - roles
      properties:
        code:
          type: string
          description: Código único da permissão
          example: read_athlete
        name:
          type: string
          description: Nome exibível da permissão
        description:
          type: string
          nullable: true
          description: Descrição da permissão
        roles:
          type: array
          items:
            type: string
          description: Papéis que possuem esta permissão

    # -------------------------------------------------------------------------
    # Membership Schemas
    # -------------------------------------------------------------------------
...
    ErrorResponse:
      type: object
      description: Resposta de erro padronizada
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Código do erro
        message:
          type: string
          description: Mensagem de erro
        details:
          type: object
          nullable: true
          description: Detalhes adicionais (regra, constraint, etc.)

  # ===========================================================================
  # Responses
  # ===========================================================================
  responses:
    UnauthorizedError:
      description: Token inválido ou ausente
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: unauthorized
            message: "Token inválido ou ausente"
            details: null
    ForbiddenError:
      description: Permissão insuficiente (R25/R26)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: permission_denied
            message: "Permissão insuficiente"
            details:
              regra: "R25"
    NotFoundError:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: not_found
            message: "Recurso não encontrado"
            details: null
    ValidationError:
      description: Erro de validação (payload inválido)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: validation_error
            message: "Erro de validação"
            details:
              errors:
                - field: "email"
                  message: "Email inválido"